# .api/apis/payclip/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Checkout Transparente",
    "description": "API para gestionar pagos y transacciones de manera transparente.",
    "contact": {
      "name": "API Support",
      "email": "developers@clip.mx"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "x-explorer-enabled": true,
  "x-samples-enabled": true,
  "x-proxy-enabled": true,
  "servers": [
    {
      "description": "URL base del ambiente de producción.",
      "url": "https://api.payclip.com"
    }
  ],
  "tags": [
    {
      "name": "Checkout Transparente",
      "description": "El recurso de **API de Checkout Transparente** ofrece una solución avanzada que te permite tener el control total de la experiencia de pago  en la cuál tus clientes pueden completar una compra sin salir de tu sitio web."
    }
  ],
  "paths": {
    "/payments": {
      "post": {
        "summary": "Realizar un pago",
        "description": "Permite realizar un cargo de manera segura y rápida. ",
        "tags": [
          "Pagos"
        ],
        "operationId": "realizarPago",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "Float",
                    "default": 1,
                    "description": "Monto del pago"
                  },
                  "currency": {
                    "type": "string",
                    "default": "MXN",
                    "description": "Moneda del pago"
                  },
                  "description": {
                    "type": "string",
                    "default": "Ejemplo de compra",
                    "description": "Descripción del pago"
                  },
                  "payment_method": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string",
                        "description": "Token del método de pago generado con el endpoint de Card Token"
                      }
                    }
                  },
                  "customer": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "default": "correo@ejemplo.com",
                        "description": "Correo electrónico del cliente"
                      },
                      "phone": {
                        "type": "string",
                        "default": "5555555555",
                        "description": "Teléfono del cliente"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El pago se ha realizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID del pago"
                    },
                    "amount": {
                      "type": "number",
                      "format": "double",
                      "description": "Monto del pago"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Objeto de error.",
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "description": "Código de error",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensaje de error.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Información detallada del error.",
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "Bad-request-1": {
                    "value": {
                      "error_code": "CL1307",
                      "message": "Bad Request",
                      "detail": "card token not found"
                    }
                  },
                  "Bad-request-2": {
                    "value": {
                      "error_code": "CL1310",
                      "message": "Bad Request",
                      "detail": "Installment not found or invalid"
                    }
                  },
                  "Bad-request-3": {
                    "value": {
                      "error_code": "CL1600",
                      "message": "Bad Request",
                      "detail": "Corrupted data: invalid emv amout, QPS international, invalid terminal country"
                    }
                  },
                  "Bad-request-4": {
                    "value": {
                      "error_code": "CL1607",
                      "message": "Bad Request",
                      "detail": "Card token expired"
                    }
                  },
                  "Bad-request-5": {
                    "value": {
                      "error_code": "CL2200",
                      "message": "Bad Request",
                      "detail": "Input validation"
                    }
                  },
                  "Bad-request-6": {
                    "value": {
                      "error_code": "CL2221",
                      "message": "Bad Request",
                      "detail": "Invalid card holder verification method"
                    }
                  },
                  "Bad-request-7": {
                    "value": {
                      "error_code": "CL1125",
                      "message": "Bad Request",
                      "detail": "Valera not allowed for this merchant"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Obtener una lista de transacciones",
        "description": "Permite obtener una lista de todos tus pagos dentro de un rango de fechas determinado.",
        "operationId": "obtenerTransacciones",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Pagos"
        ],
        "responses": {
          "200": {
            "description": "Lista de transacciones obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pages": {
                      "type": "object",
                      "properties": {
                        "size": {
                          "type": "integer"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "total_results": {
                          "type": "integer"
                        }
                      }
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "ID del pago"
                          },
                          "amount": {
                            "type": "number",
                            "format": "double",
                            "description": "Monto del pago"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Objeto de error.",
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "description": "Código de error",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensaje de error.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Información detallada del error.",
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "Bad-request-1": {
                    "value": {
                      "error_code": "BR1301",
                      "message": "Not Found",
                      "detail": "Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payments/{payment_id}": {
      "get": {
        "summary": "Obtener detalles de una transacción individual",
        "description": "Permite obtener el detalle de un pago específico.",
        "operationId": "obtenerDetalleTransaccion",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "description": "ID de la transacción",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la transacción obtenidos exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID del pago"
                    },
                    "amount": {
                      "type": "number",
                      "format": "double",
                      "description": "Monto del pago"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
      "successPaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del pago"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Monto del pago"
          }
        }
      },
      "successListResponse": {
        "type": "object",
        "properties": {
          "pages": {
            "type": "object",
            "properties": {
              "size": {
                "type": "integer"
              },
              "number": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "total_results": {
                "type": "integer"
              }
            }
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "ID del pago"
                },
                "amount": {
                  "type": "number",
                  "format": "double",
                  "description": "Monto del pago"
                }
              }
            }
          }
        }
      },
      "successDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del pago"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Monto del pago"
          }
        }
      },
      "error": {
        "description": "Objeto de error.",
        "type": "object",
        "properties": {
          "error_code": {
            "description": "Código de error",
            "type": "string"
          },
          "message": {
            "description": "Mensaje de error.",
            "type": "string"
          },
          "detail": {
            "description": "Información detallada del error.",
            "type": "string"
          }
        }
      }
    }
  }
}